Update relax-v3-working.sh
The third version of the automated DFTB+ geometry and electron optimization. This version includes user prompted input for the name of the COF, the initial SCC tolerance, the name of the geometry file, and whether or not the calculation is a restart (do you have a charges.bin file to initialize the calculation with). The script will return to the background with a no-hangup calculation after these inputs have been given by the user. This version calculates the cores to use corresponding to the number of atoms in the input geometry file. It also automatically writes the dftb_in.hsd scripts, including Hubbard derivatives and max angular momentum orbitals based on the atom types in the input geometry file. Functions have been implemented in the version to submit SCC and Forces-only calculations, as well as writing the dftb_in.hsd script for an SCC calculation and the dftb_in.hsd for a Forces-only calculation. Based on the value of the returned variable, RESULT, the next calculation in the flow chart is submitted.